pyproject.toml
[project]
name = "ragflow-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
    "uvicorn>=0.40.0",
]
